requirements.txt
python-telegram-bot[job-queue]==22.5
python-dotenv>=1.0.0
psycopg[binary]>=3.1